backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
pydantic==2.10.3
pydantic-settings==2.7.0
psycopg2-binary==2.9.10
alembic==1.14.0
pytest==8.3.5
pytest-asyncio==0.25.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20